[FIX] stock: allow re-reservation of unreserved chained move when quant slips away
Fixes odoo/odoo#6272.

Unreserved chained move can be an MTO move or a chained MTS move.

Chained MTS move is always allowed to be automatically re-reserved (by running
a scheduler or by manually checking for availability).
MTO move is only allowed to be re-reserved if user explicitly allows it: if
user decides it is OK to re-reserve this MTO move from stock and explicitly
changes move's Supply Method to MTS.

A move is recognized as an 'unreserved chained move' if it has ancestors, and
all of them are done, but the move itself is still in 'waiting' state.