[160146039] user: company-dependent signature (property field)
odoo:
- support company-dependent HTML fields
- two context.get('uid', uid) parts are needed to avoid having to
force_company into context before every write to a property field - as
in original commit #a4c7c. Having to force_company is inconvenient and
error prone as chances are you will forget to force it or will never be
able to cover all the possible usages/combinations beforehand. And you
have to effectively force_company because more often than not writing
happens as sudo(), but admin is generally not guaranteed to be in the
company of the current user, which is particularly critical for a
company-dependent field (property).